Apply isort to theano.gpuarray and remove circular references
Two new modules were made in order to remove the circular references:
`theano.gpuarray.dnn_opt` (i.e. optimizations specific to `theano.gpuarray.dnn`)
and `theano.gpuarray.optdb` (i.e. optimization databases for gpuarray `Op`s).
